+++ spec/normal_mode_code_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe "Code" do
   ]
 
   inputs.each do |input|
-    code.bind(input, num_retry: NUM_RETRY)
+    code.bind(input, num_retrials: SETTINGS["num_retrials"])
   end
 
   it "gives as many responses as the number of prompts given" do