+++ spec/reserach_mode_chat_spec.rb
@@ -15,7 +15,7 @@ RSpec.describe "MonadicApp::Chat" do
   ]
 
   inputs.each do |input|
-    chat.bind(input, num_retry: NUM_RETRY)
+    chat.bind(input, num_retrials: SETTINGS["num_retrials"])
   end
 
   chat.show_data